backend/requirements.txt
# Flask and Web Server
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-WTF==1.1.1 # Can be removed if not using Flask-Forms
Flask-Cors==4.0.0 # Added for API
PyJWT==2.8.0 # Added for token authentication

# Data and ML
pandas==2.1.1
numpy==1.26.4
ultralytics==8.3.0
opencv-python-headless==4.9.0.80
torch==2.3.0
torchvision==0.18.0
Pillow==10.0.1

# S3 and Async Tasks
boto3==1.28.63
celery==5.4.0
redis==5.0.4
eventlet==0.40.0 # FIX: Added for stable Celery execution

# Reporting (if still needed on backend)
reportlab==4.0.4